+++ src/IOEither.ts
@@ -173,6 +173,13 @@ export function getFilterable<E>(M: Monoid<E>): Filterable2C<URI, E> {
 }
 
 /**
+ * @since 2.4.0
+ */
+export function chainEither<E, A, B>(f: (a: A) => Either<E, B>): (ma: IOEither<E, A>) => IOEither<E, B> {
+  return chain(a => fromEither(f(a)))
+}
+
+/**
  * @since 2.0.0
  */
 export const ioEither: Monad2<URI> & Bifunctor2<URI> & Alt2<URI> & MonadIO2<URI> & MonadThrow2<URI> = {