+++ CHANGELOG.md
@@ -22,6 +22,19 @@ high state of flux, you're at risk of it changing without notice.
     - add `getMonoid` (@vicrac)
   - `Ordering`
     - add `monoidOrdering` (@gcanti)
+  - `TaskEither`
+    - add `chainEither`, `chainIOEither` (@gcanti)
+  - `IOEither`
+    - add `chainEither` (@gcanti)
+  - `ReaderEither`
+    - add `chainEither` (@gcanti)
+  - `ReaderTask`
+    - add `run` (@gcanti)
+    - add `chainTask` (@gcanti)
+  - `ReaderTaskEither`
+    - add `chainEither`, `chainTaskEither` (@gcanti)
+  - `StateReaderTaskEither`
+    - add `chainEither`, `chainTaskEither`, `chainReaderTaskEither` (@gcanti)
 - **Deprecation**
   - `Ord`
     - deprecate `getSemigroup` in favor of `getMonoid` (@gcanti)